refactor: navigation 함수 한 파일에서 관리
